+++ handle_upperhex.c
@@ -6,7 +6,7 @@
 /*   By: rfoo <rfoo@student.42singapore.sg>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/01/14 17:49:09 by rfoo              #+#    #+#             */
-/*   Updated: 2026/01/20 12:14:26 by rfoo             ###   ########.fr       */
+/*   Updated: 2026/01/20 16:29:21 by rfoo             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 int	handle_upperhex(va_list *args)
 {
-	unsigned long	n;
+	unsigned int	n;
 
-	n = va_arg(*args, unsigned long);
-	return (print_hex(n, 1));
+	n = va_arg(*args, unsigned int);
+	return (print_hex((unsigned long)n, 1));
 }
